Fully separated utility functions from command menu
