+++ src/commands/fun/owoify.ts
@@ -1,10 +1,10 @@
 /// @ts-nocheck
 import Command from "../../core/command";
-import {CommonLibrary, getContent} from "../../core/lib";
+import {getContent} from "../../core/libd";
 
 export default new Command({
     description: "OwO-ifies the input.",
-    async run($: CommonLibrary): Promise<any> {
+    async run($) {
         let url = new URL(`https://nekos.life/api/v2/owoify?text=${$.args.join(" ")}`);
         const content = await getContent(url.toString());
         $.channel.send(content.owo);
