+++ src/commands/utilities/shorten.ts
@@ -1,12 +1,11 @@
 import Command from "../../core/command";
-import {CommonLibrary} from "../../core/lib";
 import * as https from "https";
 
 export default new Command({
     description: "Shortens a given URL.",
     run: "Please provide a URL.",
     any: new Command({
-        async run($: CommonLibrary): Promise<any> {
+        async run($) {
             https.get("https://is.gd/create.php?format=simple&url=" + encodeURIComponent($.args[0]), function (res) {
                 var body = "";
                 res.on("data", function (chunk) {
