+++ src/core/structures.ts
@@ -1,5 +1,5 @@
 import FileManager from "./storage";
-import {select, GenericJSON, GenericStructure} from "./lib";
+import {select, GenericJSON, GenericStructure} from "./libd";
 import {watch} from "fs";
 import {Guild as DiscordGuild, Snowflake} from "discord.js";
